started refactor of view
